+++ apps/meteor/app/file-upload/client/lib/fileUploadHandler.ts
@@ -5,7 +5,9 @@ import { Tracker } from 'meteor/tracker';
 Tracker.autorun(() => {
 	const userId = Meteor.userId();
 
-	if (userId) {
+	// Check for Meteor.loggingIn to be reactive and ensure it will process only after login finishes
+	// preventing race condition setting the rc_token as null forever
+	if (userId && Meteor.loggingIn() === false) {
 		const secure = location.protocol === 'https:' ? '; secure' : '';
 
 		document.cookie = `rc_uid=${escape(userId)}; path=/${secure}`;