+++ apps/meteor/tests/end-to-end/api/livechat/04-dashboards.ts
@@ -1,4 +1,5 @@
 import { faker } from '@faker-js/faker';
+import type { Credentials } from '@rocket.chat/api-client';
 import type { ILivechatDepartment, IUser } from '@rocket.chat/core-typings';
 import { Random } from '@rocket.chat/random';
 import { expect } from 'chai';
@@ -18,11 +19,9 @@ import {
 import { createAnOnlineAgent } from '../../../data/livechat/users';
 import { sleep } from '../../../data/livechat/utils';
 import { removePermissionFromAllRoles, restorePermissionToRoles, updateSetting } from '../../../data/permissions.helper';
-import type { IUserCredentialsHeader } from '../../../data/user';
 import { IS_EE } from '../../../e2e/config/constants';
 
 describe('LIVECHAT - dashboards', function () {
-	this.retries(0);
 	// This test is expected to take more time since we're simulating real time conversations to verify analytics
 	this.timeout(60000);
 
@@ -34,7 +33,7 @@ describe('LIVECHAT - dashboards', function () {
 
 	let department: ILivechatDepartment;
 	const agents: {
-		credentials: IUserCredentialsHeader;
+		credentials: Credentials;
 		user: IUser & { username: string };
 	}[] = [];
 	let avgClosedRoomChatDuration = 0;
@@ -143,7 +142,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/conversation-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/conversation-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -151,7 +154,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an array of conversation totalizers', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/conversation-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/conversation-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -207,7 +214,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/productivity-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/productivity-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -215,7 +226,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an array of productivity totalizers', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/productivity-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/productivity-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -263,7 +278,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/chats-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/chats-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -271,7 +290,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an array of chats totalizers', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/chats-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/chats-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -323,9 +346,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(
-					api('livechat/analytics/dashboards/agents-productivity-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'),
-				)
+				.get(api('livechat/analytics/dashboards/agents-productivity-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -333,9 +358,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an array of agents productivity totalizers', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(
-					api('livechat/analytics/dashboards/agents-productivity-totalizers?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'),
-				)
+				.get(api('livechat/analytics/dashboards/agents-productivity-totalizers'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -382,7 +409,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/chats?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/chats'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -390,7 +421,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an array of productivity totalizers', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/chats?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/chats'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -431,7 +466,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/chats-per-agent?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/chats-per-agent'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -439,7 +478,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an object with open and closed chats by agent', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/chats-per-agent?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/chats-per-agent'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -532,7 +575,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/chats-per-department?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/chats-per-department'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -540,7 +587,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an object with open and closed chats by department', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/chats-per-department?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/chats-per-department'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)
@@ -577,7 +628,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an "unauthorized error" when the user does not have the necessary permission', async () => {
 			await removePermissionFromAllRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/timings?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/timings'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(403);
@@ -585,7 +640,11 @@ describe('LIVECHAT - dashboards', function () {
 		it('should return an object with open and closed chats by department', async () => {
 			await restorePermissionToRoles('view-livechat-manager');
 			await request
-				.get(api('livechat/analytics/dashboards/charts/timings?start=2019-10-25T15:08:17.248Z&end=2019-12-08T15:08:17.248Z'))
+				.get(api('livechat/analytics/dashboards/charts/timings'))
+				.query({
+					start: '2019-10-25T15:08:17.248Z',
+					end: '2019-12-08T15:08:17.248Z',
+				})
 				.set(credentials)
 				.expect('Content-Type', 'application/json')
 				.expect(200)